+++ litestar/testing/request_factory.py
@@ -43,7 +43,6 @@ def _create_default_route_handler(
 
     handler = handler_decorator("/", sync_to_thread=False, **(handler_kwargs or {}))(_default_route_handler).merge(app)
     handler._app = app
-    # handler.owner = app
     return handler
